+++ wgpu/src/api/surface.rs
@@ -235,10 +235,15 @@ static_assertions::assert_impl_all!(Surface<'_>: Send, Sync);
 
 crate::cmp::impl_eq_ord_hash_proxy!(Surface<'_> => .inner);
 
-/// Super trait for window handles as used in [`SurfaceTarget`].
-pub trait WindowHandle: HasWindowHandle + HasDisplayHandle + WasmNotSendSync {}
+/// [`Send`]/[`Sync`] blanket trait for [`HasWindowHandle`] used in [`SurfaceTarget`].
+pub trait WindowHandle: HasWindowHandle + WasmNotSendSync {}
 
-impl<T> WindowHandle for T where T: HasWindowHandle + HasDisplayHandle + WasmNotSendSync {}
+impl<T: HasWindowHandle + WasmNotSendSync> WindowHandle for T {}
+
+/// Super trait for a pair of display and window handles as used in [`SurfaceTarget`].
+pub trait DisplayAndWindowHandle: WindowHandle + HasDisplayHandle {}
+
+impl<T> DisplayAndWindowHandle for T where T: WindowHandle + HasDisplayHandle {}
 
 /// The window/canvas/surface/swap-chain/etc. a surface is attached to, for use with safe surface creation.
 ///
@@ -249,7 +254,7 @@ impl<T> WindowHandle for T where T: HasWindowHandle + HasDisplayHandle + WasmNot
 /// See also [`SurfaceTargetUnsafe`] for unsafe variants.
 #[non_exhaustive]
 pub enum SurfaceTarget<'window> {
-    /// Window handle producer.
+    /// Window and display handle producer.
     ///
     /// If the specified display and window handle are not supported by any of the backends, then the surface
     /// will not be supported by any adapters.
@@ -262,8 +267,18 @@ pub enum SurfaceTarget<'window> {
     /// # Panics
     ///
     /// - On macOS/Metal: will panic if not called on the main thread.
-    /// - On web: will panic if the `raw_window_handle` does not properly refer to a
+    /// - On web: will panic if the [`HasWindowHandle`] does not properly refer to a
     ///   canvas element.
+    /// - On all platforms: If [`crate::InstanceDescriptor::display`] was not [`None`]
+    ///   but its value is not identical to that returned by [`HasDisplayHandle::display_handle()`].
+    DisplayAndWindow(Box<dyn DisplayAndWindowHandle + 'window>),
+
+    /// Window handle producer.
+    ///
+    /// [`HasWindowHandle`]-only version of [`SurfaceTarget::DisplayAndWindow`].
+    ///
+    /// This requires that the display handle was already passed through
+    /// [`crate::InstanceDescriptor::display`].
     Window(Box<dyn WindowHandle + 'window>),
 
     /// Surface from a `web_sys::HtmlCanvasElement`.
@@ -291,12 +306,19 @@ pub enum SurfaceTarget<'window> {
     OffscreenCanvas(web_sys::OffscreenCanvas),
 }
 
+impl<'a> SurfaceTarget<'a> {
+    /// Constructor for [`Self::Window`] without consuming a display handle
+    pub fn from_window_without_display(window: impl WindowHandle + 'a) -> Self {
+        Self::Window(Box::new(window))
+    }
+}
+
 impl<'a, T> From<T> for SurfaceTarget<'a>
 where
-    T: WindowHandle + 'a,
+    T: DisplayAndWindowHandle + 'a,
 {
     fn from(window: T) -> Self {
-        Self::Window(Box::new(window))
+        Self::DisplayAndWindow(Box::new(window))
     }
 }
 
@@ -314,6 +336,9 @@ pub enum SurfaceTargetUnsafe {
     /// If the specified display and window handle are not supported by any of the backends, then the surface
     /// will not be supported by any adapters.
     ///
+    /// If the `raw_display_handle` is not [`None`] here and was not [`None`] in
+    /// [`crate::InstanceDescriptor::display`], their values _must_ be identical.
+    ///
     /// # Safety
     ///
     /// - `raw_window_handle` & `raw_display_handle` must be valid objects to create a surface upon.
@@ -321,9 +346,9 @@ pub enum SurfaceTargetUnsafe {
     ///   [`Surface`] is  dropped.
     RawHandle {
         /// Raw display handle, underlying display must outlive the surface created from this.
-        raw_display_handle: raw_window_handle::RawDisplayHandle,
+        raw_display_handle: Option<raw_window_handle::RawDisplayHandle>,
 
-        /// Raw display handle, underlying window must outlive the surface created from this.
+        /// Raw window handle, underlying window must outlive the surface created from this.
         raw_window_handle: raw_window_handle::RawWindowHandle,
     },
 
@@ -389,18 +414,38 @@ pub enum SurfaceTargetUnsafe {
 }
 
 impl SurfaceTargetUnsafe {
+    /// Creates a [`SurfaceTargetUnsafe::RawHandle`] from a display and window.
+    ///
+    /// The `display` is optional and may be omitted if it was also passed to
+    /// [`crate::InstanceDescriptor::display`].  If passed to both it must (currently) be identical.
+    ///
+    /// # Safety
+    ///
+    /// - `display` must outlive the resulting surface target
+    ///   (and subsequently the surface created for this target).
+    /// - `window` must outlive the resulting surface target
+    ///   (and subsequently the surface created for this target).
+    pub unsafe fn from_display_and_window(
+        display: &impl HasDisplayHandle,
+        window: &impl HasWindowHandle,
+    ) -> Result<Self, raw_window_handle::HandleError> {
+        Ok(Self::RawHandle {
+            raw_display_handle: Some(display.display_handle()?.as_raw()),
+            raw_window_handle: window.window_handle()?.as_raw(),
+        })
+    }
+
     /// Creates a [`SurfaceTargetUnsafe::RawHandle`] from a window.
     ///
     /// # Safety
     ///
     /// - `window` must outlive the resulting surface target
     ///   (and subsequently the surface created for this target).
-    pub unsafe fn from_window<T>(window: &T) -> Result<Self, raw_window_handle::HandleError>
-    where
-        T: HasDisplayHandle + HasWindowHandle,
-    {
+    pub unsafe fn from_window(
+        window: &impl HasWindowHandle,
+    ) -> Result<Self, raw_window_handle::HandleError> {
         Ok(Self::RawHandle {
-            raw_display_handle: window.display_handle()?.as_raw(),
+            raw_display_handle: None,
             raw_window_handle: window.window_handle()?.as_raw(),
         })
     }